requirements.txt
# python requirements file

# used by the logging-to-MQ functionality
python-crontab
paho-mqtt==1.6.0
# used by the send-to-youtube functionality
google-api-python-client
google_auth_oauthlib
ephem
paramiko
# used to force-reboot if allsky is stuck
RPi.GPIO